refactor(adapters): move component, vulnerability, property builders to builders/ (#435)
- Add builders/component.rs: build_all + private build_license
- Add builders/vulnerability.rs: build_all + private build
- Add builders/property.rs: from_resolution_guide + from_license_compliance
- Update builders/mod.rs to declare the three new submodules
- Shrink cyclonedx_formatter/mod.rs to struct, SbomFormatter impl, and module
  declarations only (~50 lines, all builder methods removed)

Closes #435
