+++ src/adapters/outbound/formatters/cyclonedx_formatter/builders/mod.rs
@@ -1,2 +1,5 @@
+pub(super) mod component;
 pub(super) mod dependency;
 pub(super) mod metadata;
+pub(super) mod property;
+pub(super) mod vulnerability;
